+++ packages/extensions/lib/extensions.dart
@@ -1,3 +1,4 @@
+export 'ai.dart';
 export 'caching.dart';
 export 'configuration.dart';
 export 'configuration_io.dart';